`sleep_fps` now requires an f32 instead of u32

+++ examples/quick-start.rs
@@ -4,7 +4,7 @@ use gemini_engine::elements::{
 };
 use gemini_engine::gameloop;
 
-const FPS: u32 = 30;
+const FPS: f32 = 30.0;
 
 fn main() {
     let mut view = View::new(40, 8, ColChar::BACKGROUND);
